Use Dead Man's Snitch to monitor `alert.yml`
Use Dead Man's Snitch (the new `snitch.yml` workflow that this commit
adds) to monitor the `alert.yml` workflow and make sure it's completing
successfully, instead of having the `slack.yml` workflow post into Slack
when `alert.yml` fails.

This is better because it will alerts us if `alert.yml` is no longer
running, not only if it's running but failing.

Dead Man's Snitch will post the alerts into Slack for us.